- general refactoring - enhanced prefix-cache-scorer cache-tracking mode's configurability - updated docs - enhanced docs (no HTML use)
