+++ .github/workflows/ci-pr-checks.yaml
@@ -25,6 +25,14 @@ jobs:
           go-version: "${{ env.GO_VERSION }}"
           cache-dependency-path: ./go.sum
 
+      - name: Install libzmq dependencies (kvcache/kvevents)
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libzmq3-dev pkg-config
+
+      - name: Set PKG_CONFIG_PATH
+        run: echo "PKG_CONFIG_PATH=/usr/lib/pkgconfig" >> $GITHUB_ENV
+
       - name: go mod tidy
         run: go mod tidy
